# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CloudSentinel API",
    "description": "\nRead-only AWS security scanning: discover resources, evaluate security rules, score risk.\n\n**Authentication.** `POST /api/auth/login` returns a short-lived access token; send it as\n`Authorization: Bearer <token>`. A refresh token is set as an httpOnly cookie and exchanged at\n`POST /api/auth/refresh`. Roles: VIEWER (read), ANALYST (+ scans and triage), ADMIN (+ users,\nAWS accounts, audit log). The role a route needs is in its description and in docs/api.md.\n\n**Lists** return one page (`limit`, `offset`); the `X-Total-Count` header holds the total.\n\n**Errors** have a `detail` field. Validation errors (422) never echo the submitted values.\n",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/scans": {
      "post": {
        "tags": [
          "scans"
        ],
        "summary": "Start Scan",
        "description": "Queues a read-only scan and returns immediately. Poll GET /api/scans/{id} for progress.",
        "operationId": "start_scan_api_scans_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScanCreate"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScanPublic"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in, or the access token expired",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          },
          "403": {
            "description": "Signed in, but your role does not allow this",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          },
          "409": {
            "description": "Conflicts with the current state (see the detail message)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "scans"
        ],
        "summary": "List Scans",
        "operationId": "list_scans_api_scans_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "aws_account_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Aws Account Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ScanPublic"
                  },
                  "title": "Response List Scans Api Scans Get"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in, or the access token expired",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          },
          "403": {
            "description": "Signed in, but your role does not allow this",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/scans/{scan_id}": {
      "get": {
        "tags": [
          "scans"
        ],
        "summary": "Get Scan",
        "operationId": "get_scan_api_scans__scan_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "scan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Scan Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScanPublic"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in, or the access token expired",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          },
          "403": {
            "description": "Signed in, but your role does not allow this",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/resources": {
      "get": {
        "tags": [
          "resources"
        ],
        "summary": "List Resources",
        "operationId": "list_resources_api_resources_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "aws_account_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Aws Account Id"
            }
          },
          {
            "name": "resource_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ResourceType"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Resource Type"
            }
          },
          {
            "name": "region",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 32
                },
                {
                  "type": "null"
                }
              ],
              "title": "Region"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ResourceSummary"
                  },
                  "title": "Response List Resources Api Resources Get"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in, or the access token expired",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/resources/{resource_uuid}": {
      "get": {
        "tags": [
          "resources"
        ],
        "summary": "Get Resource",
        "operationId": "get_resource_api_resources__resource_uuid__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "resource_uuid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Resource Uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResourceDetail"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in, or the access token expired",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorMessage": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail"
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "ErrorMessage"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ResourceDetail": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "aws_account_id": {
            "type": "string",
            "format": "uuid",
            "title": "Aws Account Id"
          },
          "region": {
            "type": "string",
            "title": "Region"
          },
          "resource_type": {
            "type": "string",
            "title": "Resource Type"
          },
          "resource_id": {
            "type": "string",
            "title": "Resource Id"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "first_seen": {
            "type": "string",
            "format": "date-time",
            "title": "First Seen"
          },
          "last_seen": {
            "type": "string",
            "format": "date-time",
            "title": "Last Seen"
          },
          "last_scan_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Scan Id"
          },
          "config": {
            "additionalProperties": true,
            "type": "object",
            "title": "Config"
          }
        },
        "type": "object",
        "required": [
          "id",
          "aws_account_id",
          "region",
          "resource_type",
          "resource_id",
          "name",
          "first_seen",
          "last_seen",
          "last_scan_id",
          "config"
        ],
        "title": "ResourceDetail"
      },
      "ResourceSummary": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "aws_account_id": {
            "type": "string",
            "format": "uuid",
            "title": "Aws Account Id"
          },
          "region": {
            "type": "string",
            "title": "Region"
          },
          "resource_type": {
            "type": "string",
            "title": "Resource Type"
          },
          "resource_id": {
            "type": "string",
            "title": "Resource Id"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "first_seen": {
            "type": "string",
            "format": "date-time",
            "title": "First Seen"
          },
          "last_seen": {
            "type": "string",
            "format": "date-time",
            "title": "Last Seen"
          },
          "last_scan_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Scan Id"
          }
        },
        "type": "object",
        "required": [
          "id",
          "aws_account_id",
          "region",
          "resource_type",
          "resource_id",
          "name",
          "first_seen",
          "last_seen",
          "last_scan_id"
        ],
        "title": "ResourceSummary"
      },
      "ResourceType": {
        "type": "string",
        "enum": [
          "AWS::Account",
          "AWS::EC2::Instance",
          "AWS::EC2::SecurityGroup",
          "AWS::S3::Bucket",
          "AWS::IAM::User",
          "AWS::IAM::Role",
          "AWS::CloudTrail::Trail"
        ],
        "title": "ResourceType"
      },
...
      "ScanCreate": {
        "properties": {
          "aws_account_id": {
            "type": "string",
            "format": "uuid",
            "title": "Aws Account Id"
          }
        },
        "type": "object",
        "required": [
          "aws_account_id"
        ],
        "title": "ScanCreate"
      },
      "ScanPublic": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "aws_account_id": {
            "type": "string",
            "format": "uuid",
            "title": "Aws Account Id"
          },
          "status": {
            "$ref": "#/components/schemas/ScanStatus"
          },
          "triggered_by_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Triggered By Id"
          },
          "caller_arn": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Caller Arn"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "resource_count": {
            "type": "integer",
            "title": "Resource Count"
          },
          "resource_counts": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Resource Counts"
          },
          "coverage": {
            "additionalProperties": true,
            "type": "object",
            "title": "Coverage"
          },
          "error_summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Summary"
          },
          "finding_count": {
            "type": "integer",
            "title": "Finding Count"
          },
          "finding_counts": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Finding Counts"
          },
          "risk_summary": {
            "additionalProperties": true,
            "type": "object",
            "title": "Risk Summary"
          },
          "rule_results": {
            "additionalProperties": true,
            "type": "object",
            "title": "Rule Results"
          }
        },
        "type": "object",
        "required": [
          "id",
          "aws_account_id",
          "status",
          "triggered_by_id",
          "caller_arn",
          "created_at",
          "started_at",
          "finished_at",
          "resource_count",
          "resource_counts",
          "coverage",
          "error_summary",
          "finding_count",
          "finding_counts",
          "risk_summary",
          "rule_results"
        ],
        "title": "ScanPublic"
      },
      "ScanStatus": {
        "type": "string",
        "enum": [
          "PENDING",
          "RUNNING",
          "COMPLETED",
          "COMPLETED_WITH_ERRORS",
          "FAILED"
        ],
        "title": "ScanStatus"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}